+++ src/main/java/domain/Count.java
@@ -1,6 +1,6 @@
 package domain;
 
-import util.validator.CountValidator;
+import exception.NegativeArgumentException;
 
 public class Count {
 
@@ -8,9 +8,36 @@ public class Count {
     private final int totalCount;
 
     public Count(String count, int totalCount) {
-        CountValidator.validate(count, totalCount);
+        validate(count, totalCount);
         this.count = Integer.parseInt(count);
         this.totalCount = totalCount;
+    }
+
+    public static void validate(String input, int totalCount) {
+        validateNumberFormat(input);
+        int count = Integer.parseInt(input);
+        validatePositive(count);
+        validateBudget(count, totalCount);
+    }
+
+    private static void validateNumberFormat(String input) {
+        try {
+            Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new NumberFormatException("숫자가 아닌 문자열이 입력되었습니다.");
+        }
+    }
+
+    private static void validatePositive(int count) {
+        if (count < 0) {
+            throw new NegativeArgumentException("개수는 항상 양수입니다.");
+        }
+    }
+
+    private static void validateBudget(int count, int totalCount) {
+        if (count > totalCount) {
+            throw new IllegalArgumentException("금액이 모자랍니다.");
+        }
     }
 
     public int getCount() {
